+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - 24/02/2021
+## [0.3.0] - COMING
 
 ### Changed
 
 - Variables are stored in one Variable registry and not one for each scope.
+- Variable scope `global` changed to `world` because global can be confusing
+- DriverManager has a DriverConfiguration to access FirefoxProfile, Server Url and Http proxy
 
 ## [0.2.0] - 23/02/2021